requirements.txt
Flask==3.0.0
pymongo==4.6.1
APScheduler==3.10.4
python-dotenv==1.0.0
Werkzeug==3.0.1
Flask-Limiter==3.5.0
bcrypt==4.1.2
gunicorn==21.2.0
requests==2.31.0
waitress==3.0.0
